requirements.txt
google-generativeai
flask
pytessaract